% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_alg_GMM.R
\name{EM_alg_GMM}
\alias{EM_alg_GMM}
\title{Robust EM Algorithm for Gaussian Mixture Models.}
\usage{
EM_alg_GMM(sampleMat, cluster, lambda = 10, inits)
}
\arguments{
\item{sampleMat}{All the points in a matrix}

\item{cluster}{Number of clusters}

\item{lambda}{Regularization Parameter}

\item{inits}{List of initial values, include:

mean (a c*d matrix)

Covariance (a list with c matrices of dimension d*d)

proportion of clusters (a vector with c elements adding up to one)}
}
\value{
A list of results, include:

Updated mean vectors for each cluster

Updated covariances for each cluster

Probability of each point in each cluster(a cluster*n matrix)

Number of clusters (cluster)

Dimension (d)

Number of points (n)
}
\description{
Given the prespecified number of clusters,
the algorithm will calculate the mean and covariance of each cluster. For complete examples
check out the documentation for robustEM.
}
