% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_hier.R
\name{initial_hier}
\alias{initial_hier}
\title{Hierarchical Clustering}
\usage{
initial_hier(sampleMat, cluster)
}
\arguments{
\item{sampleMat}{A matrix with number of rows equal to the number of points and
the number of columns equal to the dimension.}

\item{cluster}{Prespecified number of clusters}
}
\value{
A list with:

initial means (matrix with dimension c*d)

initial covariance (a list with c matrices of dimension d*d)

initial proportion of clusters tau (a vector with c elements)
}
\description{
Carry out hierarchical clustering on the raw data.
Use the outcome as the initial value for EM algorithm. For complete example please check
out the documentation for function robustEM.
}
