% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulator.R
\name{multivarGaussian}
\alias{multivarGaussian}
\title{Simulate multivariate gaussian distribution with some outliers.}
\usage{
multivarGaussian(n, d, out_perc, out_mag, cov_scale = 1)
}
\arguments{
\item{n}{Number of points}

\item{d}{Number of dimensions}

\item{out_perc}{The proportion of outliers}

\item{out_mag}{The magnitude of the outliers in terms of the covariance}

\item{cov_scale}{Constant parameter for setting the covariance of the non-outliers. Default to be 1.}
}
\value{
A list with:

mu - mean vector

sigma - covariance matrix

gauss - matrix of points
}
\description{
Simulate multivariate gaussian distribution with some outliers.
}
\examples{
simulation <- multivarGaussian(n = 200, d = 3, out_perc = 0.03, out_mag = 4)

}
