% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustEM.R
\name{robustEM}
\alias{robustEM}
\title{Robust EM Algorithm}
\usage{
robustEM(datamat, cluster, lambda = 3)
}
\arguments{
\item{datamat}{The matrix of points to be clustered.}

\item{cluster}{Number of clusters}

\item{lambda}{Regularization Parameter}
}
\value{
A list of results, include:

Updated mean vectors for each cluster

Updated covariances for each cluster

Probability of each point in each cluster(a cluster*n matrix)

Number of clusters (cluster)

Dimension (d)

Number of points (n)
}
\description{
Calculate the mean and covariance matrix of each cluster based on the data matrix.
}
\examples{
sim_info <- simMultGauss(n = 120, d = 2, cluster = 6, out_perc = 0.03, out_mag = 4)
result <- robustEM(sim_info[["simdata"]], cluster = 6)

}
