% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulator.R
\name{simMultGauss}
\alias{simMultGauss}
\title{Simulate data points from several different gaussian distributions. The number
of points from different distributions are the same.}
\usage{
simMultGauss(n, d, c, out_perc, out_mag, cov_scale = 1)
}
\arguments{
\item{n}{Number of points in each Gaussian distribution}

\item{d}{Dimension of each point}

\item{c}{Number of clusters}

\item{out_perc}{Proportion of outliers in each cluster}

\item{out_mag}{Magitude of covariance difference between outliers and non-outliers}

\item{cov_scale}{Covariance Scaling constant for the covariance of non-outliers}
}
\value{
A list with:

All the means of each cluster in a c*d matrix

All the covariances of each cluster in a list of matrices(length c, each matrix d*d)

All the simulated data points(n*c rows, d columns)
}
\description{
Simulate data points from several different gaussian distributions. The number
of points from different distributions are the same.
}
\examples{
sim_info <- simMultGauss(n = 120, d = 2, c = 6, out_perc = 0.03, out_mag = 4)


}
