% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_EM.R
\name{EM_alg_GMM}
\alias{EM_alg_GMM}
\title{Robust EM Algorithm for Gaussian Mixture Models. Given the prespecified number of clusters,
the algorithm will calculate the mean and covariance of each cluster.}
\usage{
EM_alg_GMM(sampleMat, c, lambda = 10, inits)
}
\arguments{
\item{sampleMat}{All the points in a matrix}

\item{c}{Number of clusters}

\item{lambda}{Regularization Parameter}

\item{inits}{List of initial values, include:

mean (a c*d matrix)

Covariance (a list with c matrices of dimension d*d)

proportion of clusters (a vector with c elements adding up to one)}
}
\value{
A list of results, include:

Updated mean vectors for each cluster

Updated covariances for each cluster

Probability of each point in each cluster(a c*n matrix)

Number of clusters (c)

Dimension (d)

Number of points (n)
}
\description{
Robust EM Algorithm for Gaussian Mixture Models. Given the prespecified number of clusters,
the algorithm will calculate the mean and covariance of each cluster.
}
\examples{
sim_info <- simMultGauss(n = 120, d = 2, c = 6, out_perc = 0.03, out_mag = 4)
initial_info <- initial_hier(sim_info[["simdata"]], c=6)
result <- EM_alg_GMM(sim_info[['simdata']], c=6, inits = initial_info)

}
