% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_hier.R
\name{initial_hier}
\alias{initial_hier}
\title{Carry out hierarchical clustering on the raw data.
Use the outcome as the initial value for EM algorithm.}
\usage{
initial_hier(sampleMat, c)
}
\arguments{
\item{sampleMat}{A matrix with number of rows equal to the number of points and
the number of columns equal to the dimension.}

\item{c}{Prespecified number of clusters}
}
\value{
A list with:

initial means (matrix with dimension c*d)

initial covariance (a list with c matrices of dimension d*d)

initial proportion of clusters tau (a vector with c elements)
}
\description{
Carry out hierarchical clustering on the raw data.
Use the outcome as the initial value for EM algorithm.
}
\examples{
sim_info <- simMultGauss(n = 120, d = 2, c = 6, out_perc = 0.03, out_mag = 4)
initial_info <- initial_hier(sim_info[["simdata"]], c=6)

}
